fix(quickshare): make every width conversion explicit
113 narrowing, sign and 64-to-32 warnings in QuickShare's protocol and crypto paths, now
zero. clang-tidy is at 48 warnings overall, down from 446 when the gate was first run.

Mostly `qsizetype` meeting APIs that want `int` or `size_t`: protobuf's
Serialize/ParseFromArray, `ByteSizeLong()` feeding `QByteArray::resize`, OpenSSL's
buffer lengths, `std::string` sizes reaching QByteArray and QString constructors. The
values are all bounded by frame and key sizes, so the casts change nothing at runtime --
they say out loud which width was intended, which is the point in code that parses
attacker-supplied frames.

Three worth calling out:

  * `EVP_PKEY_CTX_set1_hkdf_salt`, `_key` and `add1_hkdf_info` take **int**, not size_t.
    A first pass cast them to size_t, which silenced nothing and would have introduced
    the narrowing it was meant to remove. Read the signature, do not pattern-match the
    neighbours.
  * The auth-string hash iterated `unsigned char` over a QByteArray, which yields `char`
    -- a signedness conversion on every element, immediately undone by a cast back to
    signed char. It iterates `char` now and converts once.
  * `QuickShareDiscovery` assigned `toUInt()` straight into an int port.

Two process notes, since both cost time. `run-clang-tidy -fix` cannot be trusted
positionally: a generic fixer that wrapped expressions at the reported column produced 63
build errors, because several warnings share a line and the reported column is not always
an expression start. Pattern-by-pattern with a build between each is slower and works.
And `git checkout -- plugin/src/` to undo that experiment also reverted an earlier
uncommitted pass, which is why the count went *up* once mid-session -- commit each
verified step.

+++ plugin/src/Caelestia/Services/QuickShare/QuickShareDiscovery.cpp
@@ -140,7 +140,7 @@ bool QuickShareDiscovery::advertise(const QString& deviceName, int port) {
         static_cast<uint>(0), // flags
         serviceName, u"_FC9F5ED42C8A._tcp"_s, u"local"_s,
         QString(), // host
-        QVariant::fromValue<quint16>(port), QVariant::fromValue(txtRecord));
+        QVariant::fromValue<quint16>(static_cast<quint16>(port)), QVariant::fromValue(txtRecord));
 
     if (reply.type() == QDBusMessage::ErrorMessage) {
         qWarning() << "QuickShareDiscovery: AddService failed:" << reply.errorMessage();
@@ -248,7 +248,7 @@ void QuickShareDiscovery::onServiceResolved(const QDBusMessage& msg) {
         if (args[8].userType() == QMetaType::UShort) {
             device.port = args[8].value<quint16>();
         } else {
-            device.port = args[8].toUInt();
+            device.port = static_cast<int>(args[8].toUInt());
         }
 
         emit deviceFound(device);
